Colocando testes sobre as rotações da AVL

+++ src/tests/test_avl.cpp
@@ -182,7 +182,9 @@ void treeAvlPropreties()
             AVL::insert(tree, "b", 1);
             AVL::insert(tree, "c", 2);
             if (!(tree->root->word == "b" && tree->root->left->word == "a" && tree->root->right->word == "c" &&
-                tree->root->height == 1 && tree->root->left->height == 0 && tree->root->right->height == 0))
+                tree->root->height == 1 && tree->root->left->height == 0 && tree->root->right->height == 0) &&
+                tree->rotationsCount == 1
+            )
             {
                 AVL::destroy(tree);
                 throw std::runtime_error("A arvore nao esta sendo balanceada corretamente para o caso RR");
@@ -205,7 +207,9 @@ void treeAvlPropreties()
             AVL::insert(tree, "b", 1);
             AVL::insert(tree, "a", 2);
             if (!(tree->root->word == "b" && tree->root->left->word == "a" && tree->root->right->word == "c" &&
-                tree->root->height == 1 && tree->root->left->height == 0 && tree->root->right->height == 0))
+                tree->root->height == 1 && tree->root->left->height == 0 && tree->root->right->height == 0) &&
+                tree->rotationsCount == 1
+            )
             {
                 AVL::destroy(tree);
                 throw std::runtime_error("A arvore nao esta sendo balanceada corretamente para o caso LL");
@@ -228,7 +232,9 @@ void treeAvlPropreties()
             AVL::insert(tree, "a", 1);
             AVL::insert(tree, "b", 2);
             if (!(tree->root->word == "b" && tree->root->left->word == "a" && tree->root->right->word == "c" &&
-                tree->root->height == 1 && tree->root->left->height == 0 && tree->root->right->height == 0))
+                tree->root->height == 1 && tree->root->left->height == 0 && tree->root->right->height == 0) &&
+                tree->rotationsCount == 2
+            )
             {
                 AVL::destroy(tree);
                 throw std::runtime_error("A arvore nao esta sendo balanceada corretamente para o caso LR");
@@ -251,7 +257,9 @@ void treeAvlPropreties()
             AVL::insert(tree, "c", 1);
             AVL::insert(tree, "b", 2);
             if (!(tree->root->word == "b" && tree->root->left->word == "a" && tree->root->right->word == "c" &&
-                tree->root->height == 1 && tree->root->left->height == 0 && tree->root->right->height == 0))
+                tree->root->height == 1 && tree->root->left->height == 0 && tree->root->right->height == 0) &&
+                tree->rotationsCount == 2
+            )
             {
                 AVL::destroy(tree);
                 throw std::runtime_error("A arvore nao esta sendo balanceada corretamente para o caso RL");
